requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.68.2
google-auth==2.26.2
google-auth-oauthlib==1.2.0
gspread==5.12.4
h11==0.14.0
idna==3.6
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.13
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
starlette==0.14.2
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.15.0
